Refactor import statement in user.controllers.ts and funtions.ts
