+++ src/enums/Estados.ts
@@ -1,0 +1,6 @@
+export enum Estado {
+  Entrada = 'Entrada',
+  SalidaIntermedia = 'Salida_intermedia',
+  EntradaIntermedia = 'Entrada_intermedia',
+  Salida = 'Salida'
+}
